@@ -13,19 +13,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "^13.3.10",
-    "@angular/animations": "~13.3.12",
-    "@angular/cdk": "^13.3.9",
-    "@angular/cdk-experimental": "^13.3.9",
-    "@angular/common": "~13.3.12",
-    "@angular/compiler": "~13.3.12",
-    "@angular/core": "^13.3.12",
-    "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "~13.3.12",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~13.3.12",
-    "@angular/platform-browser-dynamic": "~13.3.12",
-    "@angular/router": "~13.3.12",
+    "@angular-devkit/core": "^14.2.10",
+    "@angular/animations": "^14.2.12",
+    "@angular/cdk": "^14.2.7",
+    "@angular/cdk-experimental": "^14.2.7",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/flex-layout": "^14.0.0-beta.41",
+    "@angular/forms": "^14.2.12",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "@ckeditor/ckeditor5-adapter-ckfinder": "^23.0.0",
     "@ckeditor/ckeditor5-angular": "^4.0.0",
     "@ckeditor/ckeditor5-autoformat": "^23.0.0",
@@ -49,8 +49,8 @@
     "@ckeditor/ckeditor5-upload": "^23.0.0",
     "@material/textfield": "^7.0.0",
     "@wiris/mathtype-ckeditor5": "^7.24.1",
-    "angular-gridster2": "^13.2.1",
-    "angular-resize-event": "^3.1.0",
+    "angular-gridster2": "^14.1.2",
+    "angular-resize-event": "^3.2.0",
     "assert": "^2.0.0",
     "browserify-fs": "^1.0.0",
     "browserify-zlib": "^0.2.0",
@@ -60,7 +60,7 @@
     "dompurify": "^2.2.0",
     "jdenticon": "^3.0.1",
     "ng2-charts": "^2.4.1",
-    "ngx-cookie-service": "^13.2.1",
+    "ngx-cookie-service": "^14.0.1",
     "ngx-jdenticon": "^1.0.3",
     "ngx-ui-scroll": "^1.11.0-rc.1",
     "node-polyfill-webpack-plugin": "^2.0.1",
@@ -80,10 +80,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "13.1.0",
-    "@angular-devkit/build-angular": "~13.3.10",
-    "@angular/cli": "~13.3.10",
-    "@angular/compiler-cli": "~13.3.12",
+    "@angular-builders/custom-webpack": "14.1.0",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
