@@ -5,7 +5,7 @@ import { LogbookInfoService } from '@shared/logbook-info.service';
 import { UserPreferencesService } from '@shared/user-preferences.service';
 import { LogbookDataService, ViewDataService } from '@shared/remote-data.service';
 import { ViewsService } from '@shared/views.service';
-import { FormBuilder } from '@angular/forms';
+import { UntypedFormBuilder } from '@angular/forms';
 import { of } from 'rxjs';
 
 
@@ -36,7 +36,7 @@ describe('ViewSettingsComponent', () => {
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       providers:[
-        FormBuilder,
+        UntypedFormBuilder,
         {provide: LogbookInfoService, useValue: logbookSpy},
         {provide: UserPreferencesService, useClass: UserPreferencesMock},
         {provide: LogbookDataService, useValue: logbookDataSpy},
